Assignment/code/requirements.txt
tensorflow-gpu==1.13.1
numpy==1.16.3
tqdm==4.31.1
matplotlib==3.0.3